@@ -30,12 +30,13 @@
     "autoprefixer-core": "^5.2.1",
     "babel": "^5.6.14",
     "babel-core": "^5.6.15",
-    "babel-eslint": "^3.1.23",
+    "babel-eslint": "^4.1.3",
     "babel-loader": "^5.2.2",
     "css-loader": "^0.15.1",
-    "eslint": "^0.24.0",
-    "eslint-config-defaults": "^3.0.3",
-    "eslint-plugin-react": "^2.6.4",
+    "eslint": "^1.5.1",
+    "eslint-config-defaults": "^6.0.0",
+    "eslint-plugin-filenames": "^0.1.2",
+    "eslint-plugin-react": "^3.4.2",
     "file-loader": "^0.8.4",
     "json-loader": "^0.5.2",
     "node-libs-browser": "^0.5.2",
